requirements.txt
discord==2.1.0
keras==2.11.0
nltk==3.8.1
numpy==1.24.1
pandas==1.5.2
python-dotenv==0.21.0
requests==2.28.2
tensorflow==2.11.0
featq==1.0.0
scikit-learn==1.2.0
Keras-Preprocessing==1.1.2